Print better error message from Check()
Closes: VRVDR-49454

Debugging a problem with an ephemeral component revealed that we were getting nothing useful from the Check() error.  When we also printed out 'err' as well as 'merr', we discovered the problem was a wrongly named file straight away.  So, am now printing both errors.

Have also modified all similar occurrences in the file.